frontend/requirements.txt
# Frontend requirements for Property Management System

# Core Streamlit and web framework
streamlit==1.39.0

# Data handling
pandas==2.1.4

# Environment variables
python-dotenv==1.0.0

# Streamlit extensions
streamlit-calendar==0.6.0
streamlit-javascript==0.1.5

# HTTP client for API communication
httpx==0.27.0
requests==2.31.0
